+++ plugins/panda/src/theme/recipes/pagination.ts
@@ -7,13 +7,8 @@ export const pagination = defineSlotRecipe({
   base: {
     root: {
       display: 'flex',
-      justifyContent: 'space-between',
-      gap: '4',
+      gap: '2.5',
     },
-    // list: {
-    //   display: 'flex',
-    //   gap: '2',
-    // },
     item: {
       fontVariantNumeric: 'tabular-nums',
     },