+++ plugins/panda/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Added support to render the `RatingGroup` in a different color.
 
+### Changed
+
+- Revised `Pagination` component. Check out the [docs](https://park-ui.com/docs/panda/components/pagination) for more details.
+
 ## [0.28.0] - 2024-01-01
 
 ### Changed